examples/detect_objects/requirements.txt
Pillow==9.1.1